requirements.txt
pyperclip~=1.8.2
python_twitter
selenium~=4.1.0
webdriver_manager
pyautogui
pyvirtualdisplay~=3.0
Pillow
pynput~=1.7.6
pytest~=7.1.1
